+++ visualisation/meshViewer.cpp
@@ -88,6 +88,8 @@ using namespace DGtal;
                                    option)
   -n [ --invertNormal ]            threshold min to define binary shape
   -v [ --drawVertex ]              draw the vertex of the mesh
+  -d [ --doSnapShotAndExit]  arg,  save display snapshot into file. Notes that the camera setting is set by default according the last saved configuration (use SHIFT+Key_M to save current camera setting in the Viewer3D). If the camera setting was not saved it will use the default camera setting.
+
  @endcode
 
 
@@ -117,12 +119,15 @@ using namespace DGtal;
 class CustomViewer3D: public Viewer3D<>
 {
 protected:
-  
+
   virtual void init()
   {
     Viewer3D<>::init();
     Viewer3D<>::setKeyDescription ( Qt::Key_I, "Display mesh informations about #faces, #vertices" );
     Viewer3D<>::setGLDoubleRenderingMode(false);
+    if(mySaveSnap){
+      QObject::connect(this, SIGNAL(drawFinished(bool)), this, SLOT(saveSnapshot(bool)));
+    }
   }
   virtual void keyPressEvent(QKeyEvent *e){
     bool handled = false;
@@ -144,6 +149,7 @@ protected:
 public: 
   std::string myInfoDisplay = "No information loaded...";
   bool myIsDisplayingInfoMode = false;
+  bool mySaveSnap = false;
 };
 
 
@@ -172,7 +178,8 @@ int main( int argc, char** argv )
   ("displaySDP,s", po::value<std::string>(), "Add the display of a set of discrete points as ball of radius 0.5.")
   ("SDPradius", po::value<double>()->default_value(0.5), "change the ball radius to display a set of discrete points (used with displaySDP option)")
   ("invertNormal,n", "invert face normal vectors." )
-  ("drawVertex,v", "draw the vertex of the mesh" );
+  ("drawVertex,v", "draw the vertex of the mesh" )
+  ("doSnapShotAndExit,d", po::value<std::string>(), "save display snapshot into file. Notes that the camera setting is set by default according the last saved configuration (use SHIFT+Key_M to save current camera setting in the Viewer3D). If the camera setting was not saved it will use the default camera setting." );
   
   bool parseOK=true;
   po::variables_map vm;
@@ -278,6 +285,10 @@ int main( int argc, char** argv )
   
   QApplication application(argc,argv);
   CustomViewer3D viewer;
+  viewer.mySaveSnap = vm.count("doSnapShotAndExit");
+  if(vm.count("doSnapShotAndExit")){
+    viewer.setSnapshotFileName(QString(vm["doSnapShotAndExit"].as<std::string>().c_str()));
+  }
   std::stringstream title;
   title  << "Simple Mesh Viewer: " << inputFilenameVect[0];
   viewer.setWindowTitle(title.str().c_str());
@@ -364,5 +375,26 @@ int main( int argc, char** argv )
   ss << "# faces: " << std::fixed << nbFaces << "    #vertex: " <<  nbVertex;
   viewer.myInfoDisplay = ss.str();
   viewer  << CustomViewer3D::updateDisplay;
+  if(vm.count("doSnapShotAndExit")){
+    // Appy cleaning just save the last snap
+    if(!viewer.restoreStateFromFile())
+      {
+        viewer.updateGL();
+      }    
+    std::string name = vm["doSnapShotAndExit"].as<std::string>();
+    std::string extension = name.substr(name.find_last_of(".") + 1);
+    std::string basename = name.substr(0, name.find_last_of("."));
+    for(int i=0; i< viewer.snapshotCounter()-1; i++){
+      std::stringstream s;
+      s << basename << "-"<< setfill('0') << setw(4)<<  i << "." << extension;
+      trace.info() << "erase temp file: " << s.str() << std::endl;
+      remove(s.str().c_str());
+    }
+    std::stringstream s;
+    s << basename << "-"<< setfill('0') << setw(4)<<  viewer.snapshotCounter()-1 << "." << extension;
+    rename(s.str().c_str(), name.c_str());
+    return 0;
+  }
+
   return application.exec();
 }